+++ docs/changelog.md
@@ -16,6 +16,7 @@ pageflow_next_text: Contributing
   - Added `DefensiveException` interface
 * Added generic exceptions that this library will want to throw
   - Added `UnsupportedType` exception
+  - Added `UnsupportedValue` exception
 * Added a way to apply a list of requirements to a piece of data
   - Added `BadRequirement` exception
   - Added `BadRequirements` exception